requirements.txt
aiohttp
beautifulsoup4
SQLAlchemy
psycopg2
Pillow